Predicting Price of a Used Toyota Car in Car Gurus data
We're using Decision tree. Problem with decision tree is overfitting, but this can be solved by pruning the branches of the tree and putting in stopping rules. This code compares performance of 3 different types of decision trees and compares their performance to a linear regression model.

What I learned---------------------------------
On average the linear regression model is $1317.74 off, ie $1317.74 off
On average the pruned tree model is 9.08% off, ie $1173 off
On average, tree model with default stopping rules is 12.3% off, ie $1618.07 off
Benchmark error was 25% or $10862 off
Out of the models tried here, the best predictor is the pruned tree model, and it surpassed the benchmark errors the most